requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
transformers>=4.30.0

# Data
kagglehub>=0.3.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=9.5.0
scikit-learn>=1.3.0

# Experiment tracking
mlflow>=2.5.0

# API
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6

# Explainability
grad-cam>=1.5.0
shap>=0.42.0

# Monitoring
prometheus-client>=0.17.0
prometheus-fastapi-instrumentator>=6.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Code quality
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
